Modify test2 fun

+++ test/mylibrary/src/main/java/com/example/mylibrary/ITest.kt
@@ -3,5 +3,5 @@ package com.example.mylibrary
 interface ITest {
     fun test(): String
     fun test1(): Int
-    fun test2(): Double
+    fun test2(): Int
 }
